Update README.md
A powerful fox known as the Nine-Tails attacks Konoha, the hidden leaf village in the Land of Fire, one of the Five Great Shinobi Countries in the Ninja World. In response, the leader of Konoha and the Fourth Hokage, Minato Namikaze seals the fox inside the body of his newborn son, Naruto Uzumaki, making Naruto a host of the beast.